Disentangle the translation config
Needed to avoid circular dependencies in the next commit.
