+++ hs-bindgen/src-internal/HsBindgen/Backend/Hs/Translation.hs
@@ -1,7 +1,6 @@
 -- | Low-level translation of the C header to a Haskell module
 module HsBindgen.Backend.Hs.Translation (
-    TranslationOpts(..)
-  , generateDeclarations
+    generateDeclarations
   ) where
 
 import Control.Arrow
@@ -28,11 +27,12 @@ import HsBindgen.Backend.Hs.Haddock.Config (HaddockConfig)
 import HsBindgen.Backend.Hs.Haddock.Documentation qualified as HsDoc
 import HsBindgen.Backend.Hs.Haddock.Translation
 import HsBindgen.Backend.Hs.Origin qualified as Origin
+import HsBindgen.Backend.Hs.Translation.Config
 import HsBindgen.Backend.SHs.AST
 import HsBindgen.Backend.SHs.AST qualified as SHs
 import HsBindgen.Backend.SHs.Translation qualified as SHs
-import HsBindgen.Backend.UniqueId
 import HsBindgen.BindingSpec qualified as BindingSpec
+import HsBindgen.Config.Internal
 import HsBindgen.Errors
 import HsBindgen.Frontend.Analysis.DeclIndex
 import HsBindgen.Frontend.Analysis.DeclIndex qualified as DeclIndex
@@ -48,69 +48,11 @@ import DeBruijn (Add (..), Env (..), Idx (..), pattern I1, pattern I2, sizeEnv,
                  tabulateEnv, weaken, zipWithEnv)
 
 {-------------------------------------------------------------------------------
-  Configuration
--------------------------------------------------------------------------------}
-
--- | Translation options
---
--- These options allow users to specify instances that /may/ be derived for all
--- structs, enums, and typedefs, along with the strategy to use.  Instances are
--- only derived when possible, however.  For example, an @Eq@ instance may only
--- be derived if all fields have @Eq@ instances.  Note that type classes that
--- @hs-bindgen@ generates instances for must not be included in this
--- configuration.
-data TranslationOpts = TranslationOpts {
-      -- | Default set of classes to derive for structs
-      translationDeriveStruct :: [(Hs.Strategy Hs.HsType, Hs.TypeClass)]
-
-      -- | Default set of classes to derive for enums
-    , translationDeriveEnum :: [(Hs.Strategy Hs.HsType, Hs.TypeClass)]
-
-      -- | Default set of classes to derive for typedefs
-    , translationDeriveTypedef :: [(Hs.Strategy Hs.HsType, Hs.TypeClass)]
-
-      -- | Ensure that identifier generated by @hs-bindgen@ are unique.
-    , translationUniqueId :: UniqueId
-    }
-  deriving stock (Show, Eq, Generic)
-
-instance Default TranslationOpts where
-  def = TranslationOpts {
-      translationDeriveStruct = [
-          (Hs.DeriveStock, Hs.Show)
-        , (Hs.DeriveStock, Hs.Eq)
-        ]
-    , translationDeriveEnum = [
-          (Hs.DeriveStock, Hs.Eq)
-        , (Hs.DeriveStock, Hs.Ord)
-        ]
-    , translationDeriveTypedef = [
-          (Hs.DeriveStock, Hs.Eq)
-        , (Hs.DeriveStock, Hs.Ord)
-        , (Hs.DeriveStock, Hs.Read)
-        , (Hs.DeriveStock, Hs.Show)
-        , (Hs.DeriveNewtype, Hs.Enum)
-        , (Hs.DeriveNewtype, Hs.Ix)
-        , (Hs.DeriveNewtype, Hs.Bounded)
-        , (Hs.DeriveNewtype, Hs.Bits)
-        , (Hs.DeriveNewtype, Hs.FiniteBits)
-        , (Hs.DeriveNewtype, Hs.Floating)
-        , (Hs.DeriveNewtype, Hs.Fractional)
-        , (Hs.DeriveNewtype, Hs.Integral)
-        , (Hs.DeriveNewtype, Hs.Num)
-        , (Hs.DeriveNewtype, Hs.Real)
-        , (Hs.DeriveNewtype, Hs.RealFloat)
-        , (Hs.DeriveNewtype, Hs.RealFrac)
-        ]
-    , translationUniqueId = def
-    }
-
-{-------------------------------------------------------------------------------
   Top-level
 -------------------------------------------------------------------------------}
 
 generateDeclarations ::
-     TranslationOpts
+     TranslationConfig
   -> HaddockConfig
   -> Hs.ModuleName
   -> DeclIndex
@@ -137,7 +79,7 @@ data WithCategory a = WithCategory {
   } deriving (Show)
 
 generateDeclarations' ::
-     TranslationOpts
+     TranslationConfig
   -> HaddockConfig
   -> Hs.ModuleName
   -> DeclIndex
@@ -388,7 +330,7 @@ getInstances instanceMap name = aux
 -- TODO: Take DeclSpec into account
 generateDecs ::
      State.MonadState InstanceMap m
-  => TranslationOpts
+  => TranslationConfig
   -> HaddockConfig
   -> Hs.ModuleName
   -> C.Decl
@@ -443,7 +385,7 @@ reifyStructFields struct k = Vec.reifyList (C.structFields struct) k
 -- | Generate declarations for given C struct
 structDecs :: forall n m.
      (SNatI n, State.MonadState InstanceMap m)
-  => TranslationOpts
+  => TranslationConfig
   -> HaddockConfig
   -> C.DeclInfo
   -> C.Struct
@@ -864,7 +806,7 @@ unionFieldDecls unionName f = [
 
 enumDecs ::
      State.MonadState InstanceMap m
-  => TranslationOpts
+  => TranslationConfig
   -> HaddockConfig
   -> C.DeclInfo
   -> C.Enum
@@ -1008,7 +950,7 @@ enumDecs opts haddockConfig info e spec = do
 
 typedefDecs ::
      State.MonadState InstanceMap m
-  => TranslationOpts
+  => TranslationConfig
   -> HaddockConfig
   -> C.DeclInfo
   -> C.Typedef
@@ -1225,7 +1167,7 @@ typedefFieldDecls hsNewType = [
 
 macroDecs ::
      State.MonadState InstanceMap m
-  => TranslationOpts
+  => TranslationConfig
   -> HaddockConfig
   -> C.DeclInfo
   -> C.CheckedMacro
@@ -1238,7 +1180,7 @@ macroDecs opts haddockConfig info checkedMacro spec =
 
 macroDecsTypedef ::
      State.MonadState InstanceMap m
-  => TranslationOpts
+  => TranslationConfig
   -> HaddockConfig
   -> C.DeclInfo
   -> C.CheckedMacroType
@@ -1644,7 +1586,7 @@ genNameFromArgs args' res' suffix =
 functionDecs ::
      HasCallStack
   => SHs.Safety
-  -> TranslationOpts
+  -> TranslationConfig
   -> HaddockConfig
   -> Hs.ModuleName
   -> C.DeclInfo
@@ -1906,7 +1848,7 @@ getMainHashIncludeArg declInfo = case C.declHeaderInfo declInfo of
 -- returns precisely the value of the constant rather than a /pointer/ to
 -- the value.
 global ::
-     TranslationOpts
+     TranslationConfig
   -> HaddockConfig
   -> Hs.ModuleName
   -> InstanceMap
@@ -1993,7 +1935,7 @@ constGetter ty instsMap info pureStubName = concat [
 --
 -- * @pureStubName@: the identifier of the /pure/ stub.
 addressStubDecs ::
-     TranslationOpts
+     TranslationConfig
   -> HaddockConfig
   -> Hs.ModuleName
   -> C.DeclInfo -- ^ The given declaration
