+++ hs-bindgen/src-internal/HsBindgen/Backend/Hs/Translation/Config.hs
@@ -1,0 +1,65 @@
+module HsBindgen.Backend.Hs.Translation.Config (
+    TranslationConfig(..)
+  ) where
+
+import HsBindgen.Backend.Hs.AST qualified as Hs
+import HsBindgen.Config.Prelims
+import HsBindgen.Imports
+
+{-------------------------------------------------------------------------------
+  Configuration
+-------------------------------------------------------------------------------}
+
+-- | Translation options
+--
+-- These options allow users to specify instances that /may/ be derived for all
+-- structs, enums, and typedefs, along with the strategy to use.  Instances are
+-- only derived when possible, however.  For example, an @Eq@ instance may only
+-- be derived if all fields have @Eq@ instances.  Note that type classes that
+-- @hs-bindgen@ generates instances for must not be included in this
+-- configuration.
+data TranslationConfig = TranslationConfig {
+      -- | Default set of classes to derive for structs
+      translationDeriveStruct :: [(Hs.Strategy Hs.HsType, Hs.TypeClass)]
+
+      -- | Default set of classes to derive for enums
+    , translationDeriveEnum :: [(Hs.Strategy Hs.HsType, Hs.TypeClass)]
+
+      -- | Default set of classes to derive for typedefs
+    , translationDeriveTypedef :: [(Hs.Strategy Hs.HsType, Hs.TypeClass)]
+
+      -- | Ensure that identifier generated by @hs-bindgen@ are unique.
+    , translationUniqueId :: UniqueId
+    }
+  deriving stock (Show, Eq, Generic)
+
+instance Default TranslationConfig where
+  def = TranslationConfig {
+      translationDeriveStruct = [
+          (Hs.DeriveStock, Hs.Show)
+        , (Hs.DeriveStock, Hs.Eq)
+        ]
+    , translationDeriveEnum = [
+          (Hs.DeriveStock, Hs.Eq)
+        , (Hs.DeriveStock, Hs.Ord)
+        ]
+    , translationDeriveTypedef = [
+          (Hs.DeriveStock, Hs.Eq)
+        , (Hs.DeriveStock, Hs.Ord)
+        , (Hs.DeriveStock, Hs.Read)
+        , (Hs.DeriveStock, Hs.Show)
+        , (Hs.DeriveNewtype, Hs.Enum)
+        , (Hs.DeriveNewtype, Hs.Ix)
+        , (Hs.DeriveNewtype, Hs.Bounded)
+        , (Hs.DeriveNewtype, Hs.Bits)
+        , (Hs.DeriveNewtype, Hs.FiniteBits)
+        , (Hs.DeriveNewtype, Hs.Floating)
+        , (Hs.DeriveNewtype, Hs.Fractional)
+        , (Hs.DeriveNewtype, Hs.Integral)
+        , (Hs.DeriveNewtype, Hs.Num)
+        , (Hs.DeriveNewtype, Hs.Real)
+        , (Hs.DeriveNewtype, Hs.RealFloat)
+        , (Hs.DeriveNewtype, Hs.RealFrac)
+        ]
+    , translationUniqueId = def
+    }
