+++ hs-bindgen/src-internal/HsBindgen/Config/Prelims.hs
@@ -1,5 +1,9 @@
-module HsBindgen.Backend.UniqueId
-  ( UniqueId (..)
+module HsBindgen.Config.Prelims (
+    -- * Base module name
+    defBaseModuleName
+
+    -- * Unique IDs
+  , UniqueId (..)
   , UniqueIdMsg (..)
   , checkUniqueId
   ) where
@@ -7,7 +11,19 @@ module HsBindgen.Backend.UniqueId
 import Text.SimplePrettyPrint qualified as PP
 
 import HsBindgen.Imports
+import HsBindgen.Language.Haskell qualified as Hs
 import HsBindgen.Util.Tracer
+
+{-------------------------------------------------------------------------------
+  Base module name
+-------------------------------------------------------------------------------}
+
+defBaseModuleName :: Hs.ModuleName
+defBaseModuleName = "Generated"
+
+{-------------------------------------------------------------------------------
+  Unique IDs
+-------------------------------------------------------------------------------}
 
 -- | C uses a global namespace. We must ensure that identifiers generated by
 -- @hs-bindgen@ are unique.
